# Make the tests run faster by reducing PREC