This speeds up the `test_series_generic` test in Debug mode from 1.09s to 0.83s.
Before:
```
$ ctest -R test_series_generic
Test project /home/certik/repos/symengine
    Start 6: test_series_generic
1/1 Test #6: test_series_generic ..............   Passed    1.09 sec

100% tests passed, 0 tests failed out of 1

Total Test time (real) =   1.10 sec
```
After:
```
$ ctest -R test_series_generic
Test project /home/certik/repos/symengine
    Start 6: test_series_generic
1/1 Test #6: test_series_generic ..............   Passed    0.83 sec

100% tests passed, 0 tests failed out of 1

Total Test time (real) =   0.84 sec
```

This helps with #1012, but still not fixed.