@certik 2019-02-18
I rebased on top of the latest master. The tests should pass now.