@isuruf 2019-04-11
I don't think the test time is a big issue. Were large prec used to test some specific code paths?