@certik 2019-04-11
> Were large prec used to test some specific code paths?

That's what I am not sure --- it doesn't seem so. And if it doesn't test anything useful, I would argue the test should be as fast as possible. As we will keep adding more tests in the future, every bit counts.